split brandings to config + scss edit in react
