+++ generators/angular/angular.js
@@ -5,7 +5,8 @@ const { updateLog, errorLog } = require('../../utils/log');
 const { deleteFolderSync } = require('../../utils/delete');
 const { updatePackageJson } = require('../../utils/package');
 const { copyFolderRecursiveSync, copyFileSync } = require('../../utils/copy');
-const { mapBranding, brandings } = require('../../utils/branding');
+const brandings = require('../../config/brandings.config');
+const { mapBranding } = require('../../utils/branding');
 const frontEndConfig = require('../../config/front-end.config');
 const { execPromise } = require('../../utils/exec');
 const { mapRouting } = require('./routing');
